Add ability for unattended creation of base and kiosk images
Also separates the Debian base image from a kiosk setup. The debian base
image is the minimum installed with some documentation removed, with
ispell removed and with nano replaced by vi. The manpages are kept
because they were added back in when installing other packages anyway.

The kiosk image adds the X server, Firefox ESR and wm2, and configures
the system in a way that automatically spawns the X server with no boot
menu.